modularized twoFrameBrowser
BrowserWindow.css describes the style of the
browser window
BrowserWindow.js declares a class BrowserWindow and some
helper functions that take care of the client side code
and we added a template tag that will set up the html and a little
bit of javascript needed to actually create the window. Inheriting from
twoFrameBrowser takes care of the necessary js and css inclusions.