added function to calculate center and fwhm of peak
NOTE: the fwhm is very dependent on the bin sizes,
and if the bins are too small, then the data is two noisy
and we get multiple zeros. I don't know what to do about this,
should we try and fit a gaussion, or a poisson distribution thing?
I just don't know what to do right now.